Overhauled the config system, finished config...
... from YAML. This allows for config.py to be edited from a simple
experiments.yaml (example located in examples/). More to come on this as
the other features intended for this come online.
