+++ odorsampling/utils.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import threading
 from contextlib import contextmanager
+import logging
 
 import numpy as np
 
@@ -52,6 +53,27 @@ def choice_gauss_activation(mu, sigma, **_):
 def expovar_activation(lambd, **_):
     return RNG.exponential(1/lambd)
 
+
+LOG_FORMATTER = logging.Formatter(config.LOG_MSG_FMT, config.LOG_DATE_FMT)
+
+def default_log_setup(logger: logging.Logger, log_level: int = None, stream_handler_level = logging.WARNING, file_handler_level = logging.DEBUG):
+    """
+    Automatically adds LOG_FILE_HANDLER and LOG_STREAM_HANDLER as handlers.
+
+    Parameters
+    ----------
+    logger -
+        The logger to setup with the default configuration.
+    """
+    logger.setLevel(config.LOG_LEVEL if log_level is None else log_level)
+    file_handler = logging.FileHandler(config.LOG_FILE_NAME)
+    stream_handler = logging.StreamHandler()
+    file_handler.setFormatter(LOG_FORMATTER)
+    stream_handler.setFormatter(LOG_FORMATTER)
+    file_handler.setLevel(file_handler_level)
+    stream_handler.setLevel(stream_handler_level)
+    logger.addHandler(file_handler)
+    logger.addHandler(stream_handler)
 
 
 class ReaderWriterSuite:
