+++ requirements.txt
@@ -2,3 +2,4 @@ numpy
 matplotlib
 scipy
 pandas
+pyyaml
